refactor: move API calls to separate module
