+++ src/components/errorBoundary/index.ts
@@ -1,0 +1,3 @@
+import ErrorBoundary from "./ErrorBoundary";
+
+export { ErrorBoundary };
